+++ docs/CHANGES.md
@@ -367,3 +367,4 @@
 
 * v1.0.0 (2016-04-11)
    * Initial public release
+