+++ docs/CHANGES.md
@@ -13,7 +13,8 @@
     as empty JSON Object (`{}`) instead of `NULL`.
   * Fixed floats being rejected as direct primitive payload.
   * Fixed `Converter` unit tests for primitives-as-payload.
-  * Updated and corrected PHPDocs (incl. some type hints)
+  * Fixed `ResponseBuilderProvider` throwing incorrect Exception in case of invalid config file.
+  * Updated and corrected PHPDocs (incl. some type hints).
   * Added missing type hint to `success()`.
   * Added strict type header to classes.
   * Updated documentation.