+++ docs/CHANGES.md
@@ -9,8 +9,8 @@
 ## CHANGE LOG ##
 
 * v9.3.0 (2021-06-19)
-  * Added `data_always_object` config option, that enforces response `data` node to be returned
-    as empty JSON Object (`{}`) instead of `NULL`.
+  * Added `data_always_object` config option that, when enabled enforces response `data` node
+    to always be JSON object (for `NULL` it will return empty object `{}`).
   * Fixed floats being rejected as direct primitive payload.
   * Fixed `Converter` unit tests for primitives-as-payload.
   * Fixed `ResponseBuilderProvider` throwing incorrect Exception in case of invalid config file.