+++ docs/CHANGES.md
@@ -11,6 +11,7 @@
 * v9.3.0 (2021-06-19)
   * Added `data_always_object` config option that, when enabled enforces response `data` node
     to always be JSON object (for `NULL` it will return empty object `{}`).
+  * Improved code quality (fully pass on PHPStan's strict mode)
   * Fixed floats being rejected as direct primitive payload.
   * Fixed `Converter` unit tests for primitives-as-payload.
   * Fixed `ResponseBuilderProvider` throwing incorrect Exception in case of invalid config file.