+++ docs/CHANGES.md
@@ -8,7 +8,7 @@
 
 ## CHANGE LOG ##
 
-* v9.3.0 (2021-06-19)
+* v9.3.0 (2021-06-21)
   * Added `data_always_object` config option that, when enabled enforces response `data` node
     to always be JSON object (for `NULL` it will return empty object `{}`).
   * Improved code quality (fully pass on PHPStan's strict mode)